+++ maps/scenarios/waterfall.py
@@ -41,7 +41,7 @@ class Scenario(BaseScenario):
             agent.set_pos(
                 torch.tensor(
                     [-0.2 + 0.1 * i, 1.0],
-                    dtype=torch.float64,
+                    dtype=torch.float32,
                     device=self.world.device,
                 ),
                 batch_index=env_index,
@@ -50,7 +50,7 @@ class Scenario(BaseScenario):
             landmark.set_pos(
                 torch.tensor(
                     [0.2 if i % 2 else -0.2, 0.6 - 0.3 * i],
-                    dtype=torch.float64,
+                    dtype=torch.float32,
                     device=self.world.device,
                 ),
                 batch_index=env_index,
@@ -58,7 +58,7 @@ class Scenario(BaseScenario):
             landmark.set_rot(
                 torch.tensor(
                     [torch.pi / 4 if i % 2 else -torch.pi / 4],
-                    dtype=torch.float64,
+                    dtype=torch.float32,
                     device=self.world.device,
                 ),
                 batch_index=env_index,
@@ -67,7 +67,7 @@ class Scenario(BaseScenario):
         floor.set_pos(
             torch.tensor(
                 [0, -1],
-                dtype=torch.float64,
+                dtype=torch.float32,
                 device=self.world.device,
             ),
             batch_index=env_index,
